requirements.txt
# AI Code Assistant — Local-only (Ollama)
# Core dependencies
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
python-multipart==0.0.12
rich==13.7.1
httpx==0.27.2

# Local AI backend (Ollama)
ollama>=0.4.0
